modularised competitive dashboard component + updated cline rules
